Moving InitializationEvents to the common api
